Add average citations statbox
Add StatBox to the publications section listing the average citations
per article.

Closes #58
